Move most of `base::types::pretty_print::Printer` functions to `format::pretty_print::Printer`
