+++ format/Cargo.toml
@@ -14,6 +14,8 @@ documentation = "https://docs.rs/gluon"
 env_logger = { version = "0.3.4", optional = true }
 log = "0.3.6"
 pretty = "0.3.2"
+itertools = "0.5.6"
+
 gluon_base = { path = "../base", version = "0.5.0" } # GLUON
 gluon_parser = { path = "../parser", version = "0.5.0" } # GLUON
